refactor: 강좌 시청 페이지 타입 및 서비스 구조 개편
- enrollment / enrolledCourse / progress 타입을 course 도메인으로 통합
- learn.ts 타입 정의 추가 및 기존 타입 정리
- CourseLearnClientPage 및 useCourseLearn 훅 로직 개선
- learnService에 API 목업 기반 데이터 페칭 로직 정리
- 중복 타입 파일(enrollment 도메인) 제거 및 구조 일원화
