+++ src/domains/course/types/enrolledCourse.ts
@@ -1,5 +1,5 @@
-import type { Enrollment } from './enrollment';
 import type { Course } from '@/domains/course/types/course';
+import { Enrollment } from './enrollment';
 
 export type EnrolledCourse = Enrollment & {
   course: Pick<Course, 'id' | 'title' | 'category' | 'thumbnailUrl'> | null;
