+++ src/domains/course/types/learn.ts
@@ -1,9 +1,9 @@
-import type { Enrollment } from '@/domains/enrollment/types/enrollment';
-import type { Progress } from '@/domains/enrollment/types/progress';
+import type { Enrollment } from '@/domains/course/types/enrollment';
+import type { Progress } from '@/domains/course/types/progress';
 
 export type LearnLectureResource = {
   resourceId: number;
-  resourceType: 'VIDEO' | 'PDF';
+  resourceType: 'VIDEO' | 'PDF' | 'ZIP' | 'DOC';
   fileUrl: string;
   isDownloadable: boolean;
 };
@@ -49,4 +49,62 @@ export type LearnPageData = {
   course: LearnCourse;
   enrollment: Enrollment | null;
   progress: Progress | null;
+};
+
+export type CourseDetailResponse = {
+  status: string;
+  code: string;
+  message: string;
+  data: LearnCourse;
+};
+
+export type EnrollmentListItem = LearnPageData['enrollment'] & {
+  enrollmentId: number;
+  courseId: number;
+};
+
+export type EnrollmentListResponse = {
+  status: string;
+  code: string;
+  message: string;
+  data: {
+    content: EnrollmentListItem[];
+    totalElements: number;
+    totalPages: number;
+    pageNumber: number;
+    pageSize: number;
+  };
+};
+
+export type ProgressResponse = {
+  status: string;
+  code: string;
+  message: string;
+  data: LearnPageData['progress'];
+};
+
+export type UILecture = {
+  id: string;
+  title: string;
+  type: 'VIDEO' | 'PDF' | 'ZIP' | 'DOC';
+  duration?: string;
+  description: string;
+  completed: boolean;
+  videoUrl?: string;
+  pdfUrl?: string;
+};
+
+export type UISection = {
+  id: string;
+  title: string;
+  description: string;
+  lectures: UILecture[];
+};
+
+export type UICourse = {
+  id: string;
+  title: string;
+  instructor: string;
+  description: string;
+  sections: UISection[];
 };
