+++ src/domains/course/types/progress.ts
