--- src/domains/enrollment/types/enrollment.ts
@@ -1,8 +1,0 @@
-export type Enrollment = {
-  id: string;
-  userId: string;
-  courseId: string;
-  progress: number;
-  enrolledAt: string;
-  updatedAt?: string;
-};
